Rename methods/vars for clarity
There are four unique structures that need to be more explicitely differentiated in dcnm_vrf

playbooks (want) - conform to playbook models
diffs - partially conform to playbook models, but not quite
payloads - conform to one of the payload models
controller responses (have, and other) - conform to controller response models.

dcnm_vrf has, from the beginning, been ambiguous with some of these as to which is which (especially payloads and diffs).  It also converts between them unnecessarily, and carries parts of playbooks within payloads for convenience (I think?)

Anyway, this module would benefit if we explicitely separate these structures and handle them separately.

To that end, and as just a small first step, we are renaming the following:

vars:

- self.want_create_models -> self.want_create_payload_models

methods:

- populate_want_create_models -> populate_want_create_payload_models