Fix update_attach_params_extension_values
1. plugins_module_utils/vrf/dcnm_vrf_v12.py

1a. update_attach_params_extension_values

In the vrf-lite model, dot1q is an int, which can be zero (0).  If this is assigned in the payload, the VRF never achieves DEPLOYED state.

We’ve added a temporary fix in this method, but should handle this in the model and I’ve added a TODO to that effect.

Committing this to verify this is the actual issue (i.e. that integration tests pass).  If so, will look into a model fix.