Move dot1q validation to model
1. plugins_module_utils/vrf/dcnm_vrf_v12.py

1a. update_attach_params_extension_values

In the previous commit, we validated the value of dot1a in this method.

We’ve moved this to a field_validator in PlaybookVrfLiteModel.

2. plugins_module_utils/vrf/model_playbook_vrf_v12.py

2a. PlaybookVrfLiteModel

Validate dot1q and serialize to a string

2b. Replace all model_validator with field_validator

I initially used model_validator out of ignorance.  field_validator is more appropriate for the validations used in this model.