Experimental fix for 500 controller response
The field serializer for extension_values in PayloadVrfsAttachmentsLanAttachListItem needs to set the value to “” if MULTISOTE_CONN and VRF_LITE_CONN both have zero length, else the controller chokes on the payload.

This commit tests one way to accomplish this.  If this works against the contoller, I’ll clean things up in the next commit.