Fix for attachment payload via model, more…
1. plugins/module_utils/vrf/transmute_diff_attach_to_payload.py

1a. self._payload is a list, but was type-hinted as a str.

1b commit()

- When populating PayloadVrfsAttachments, set vrfName to “” if it’s None.

1c. Improve logging messages.

1d. Move assignment of vrf_lite_conn_list closer to where it is used.

2. tests/unit/modules/dcnm/test_dcnm_vrf_12.py

2a. test_dcnm_vrf_12_query_vrf_lite

- Update to assert on new value for extensionValues

2b. test_dcnm_vrf_12_query_lite_without_config

- Update to assert on new value for extensionValues

3. plugins/module_utils/vrf/model_controller_response_vrfs_switches_v12.py

3a. Make most fields optional with a default so that model_construct() can build a default representation of the model.

3b. VrfLiteConnOuterItem

Sort of a hack, but now that we are returning a fully-formed model with default values for things not present in the switch response, we need a way to determine if interface(s) on the switch are VRF_LITE capable.  Previously, we did this by comparing if extensionValues was None or “”.  Now, we’ve set AUTO_VRF_LITE_FLAG to “NA” by default.  So, if an interface is not VRF_LITE capable, this will now be “NA”.

4. plugins/module_utils/vrf/dcnm_vrf_v12.py

4a. _update_vrf_lite_extension_model

Per 3b above, modify the conditional to inspect AUTO_VRF_LITE_FLAG for “NA” and skip extensionValues processing if this is the case.

5. plugins/module_utils/vrf/model_controller_response_vrfs_deployments_v12.py

5a. Make REQUEST_PATH optional with a default of “”