DcnmNetwork: Fix for overwriting TOR ports
atch_tor_ports is being initialized on each iteration of the inner for loop within the nested for loops in method diff_for_attach_deploy.

The result being that only the last TOR port is deployed.

This commit moves initialization of atch_tor-ports to the top of the method so that it is not overwritten in the inner for loop.