dcnm_network: Fix TOR port handling (#520)
* DcnmNetwork: Fix for overwriting TOR ports

atch_tor_ports is being initialized on each iteration of the inner for loop within the nested for loops in method diff_for_attach_deploy.

The result being that only the last TOR port is deployed.

This commit moves initialization of atch_tor-ports to the top of the method so that it is not overwritten in the inner for loop.

* Fix overwritting torPorts in multiple locations

1. Revert the original “fix” which wasn’t a fix.

2. The attach.update() for torPorts in 4 locations was overwritting the “torPorts” key in the dictionary rather than appending to it.  The fix is to append to a list first, and then set torPorts to the joined list.