OneManage endpoints: initial commit
# Description

This commit contains an initial implementation for onemanage/fabrics/<fabricName> endpoint.

Rather than use the existing inheritance-based endpoint structure, we’ve implemented a composition-based alternative.

Several markdown files are included that describe the differences in approach, the benefits of the new approach, a migration guide for existing endpoints, and specifics of the design.  These can be used later if/when we decide to migrate the existing endpoints.

As a summary, the new approach supports composable query params, including endpoint-specific query params, and Lucene filter query params.

## Markdown files

- common/api/QUERY_PARAMS_DESIGN.md
- common/api/PROPERTY_STYLE_COMPARISONS.md
- common/api/MIGRATION_EXAMPLE.md

## Python files

- common/api/base_paths.py
  - BasePath - centralized base paths

- common/api/query_params.py
  - QueryParams - abstract base class
  - EndpointQueryParams - endpoint-specific query parameters
  - LuceneQueryParams - Lucene-Style Query Parameters

- common/api/onemanage/endpoints.py
  - EpOneManageFabricDetails - Fabric Details Endpoint (OneManage)