Fix KeyErrors due to unprotected accesses
There are many more places in this module where the possibility for KeyError exist.  The fixes in this commit are just enough to allow the user to move forward.

Need to discuss with Mike the scope of this PR.  Should we fix all instances where KeyError is possible, or limit the fix to the places in this commit?