FabricCommon: Move rest_send,results to children
1. FabricCommon

Move results property to each class that inherits from FabricCommon, and remove it from FabricCommon.

This is so that results can be the single place where action and operation_type are set.  Previously, these where set in two places in each class.

We are also moving the rest_send property to each class that inherits from FabricCommon and removing it from FabricCommon since, in general, we are trying to lessen reliance on inheritance in favor of compostion.