VerifyPlaybookParams: replace eval() with explicit comparisons (#546)
* VerifyPlaybookParams: replace eval() with explicit comparisons

1. VerifyPlaybookParams.eval_parameter_rule()

While the usage of eval() in this case was not dangerous, it is a bad look when security scanners flag this.

This commit replaces eval() with explicit comparisons.

* VerifyPlaybookParams: remove references to eval()

1. Remove references in comments to eval()
2. Fix a few bad type annotations

* VerifyPlaybookParams: Appease black linter

Sorry, this commit contains a lot of linter changes.

But, there are no functional changes, so you can ignore during review.

* dcnm_update_arg_specs: implement without eval()

Rewrite dcnm_update_arg_specs() without eval.

I’ve tested this against the original function using a test script that works with all of the data passed through these functions by the unit tests for dcnm_vpc_pairs.

* Revert "dcnm_update_arg_specs: implement without eval()"

This reverts commit 8004586b700366822dd3817c481279d46c67b807.

Accidently committed the fix for dcnm.py into this branch.  Reverting…