Testing removal of import guard from main module.
Copilot is complaining that there are unused imports (which there are!).  My reasoning here was that Ansible needs to introspect on the imports for Pydantic, but maybe it doesn’t need this in the main module (in this case dcnm_fabric_group.py).

Let’s see if  Ansible sanity passes after removing them.  If so, this is a small win!  If not, I’ll revert this commit.