Fix for query-state config
# Summary

Fix issue where query-state configuration is ignored.

Not sure why this ever worked?

I think we initially wanted an empty query config to return all fabrics.  Maybe we still want this; in which case we need to modify query.py accordingly.

Anyway, for now, we populate self.config for query-state if params is not None.

An alternative fix would be to mandate that params is not empty and add query-state to the list of states that requires a config.

Currently, in query.py, a config is required so an empty config will result in an error here.

This fix at least ensures that a populated query config is accepted in dcnm_fabric.py and passed downstream to query.py.